+++ CHANGES.rst
@@ -4,6 +4,12 @@ Changelog
 1.0dev (unreleased)
 -------------------
 
+- Fix icalendar importer to support multiple-line EXDATE/RDATE definitions.
+  [thet]
+
+- Fix runtime error in icalendar importer.
+  [thet]
+
 - For the setup's tests extra, depend on plone.app.testing <= 4.2.2 until the
   Dexterity and Archetypes tests are split up and the tests don't have a hard
   dependency on Archetypes.